requirements.txt
Django>=3.2,<4.0
djangorestframework
gunicorn>=20.1.0

# Async tasks
#celery
#redis

# Notifications
slack_sdk

# Configuration
django-environ

# Testing
pytest
pytest-django
dependency-injector

# Environment
python-dotenv

psycopg2-binary>=2.9.1
